application/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
absl-py==0.9.0            # via tensorboard, tensorflow
aiofiles==0.4.0           # via -r requirements.in, fastapi
alembic==1.4.2            # via -r requirements.in
aniso8601==7.0.0          # via graphene
astor==0.8.1              # via tensorflow
async-exit-stack==1.0.1   # via fastapi
async-generator==1.10     # via fastapi
bcrypt==3.1.7             # via passlib
certifi==2020.4.5.1       # via requests
cffi==1.14.0              # via bcrypt
chardet==3.0.4            # via requests
click==7.1.1              # via uvicorn
decorator==4.4.2          # via networkx
dnspython==1.16.0         # via email-validator
email-validator==1.0.5    # via fastapi
fastapi[all]==0.54.0      # via -r requirements.in
gast==0.3.3               # via tensorflow
google-pasta==0.2.0       # via tensorflow
graphene==2.1.8           # via fastapi
graphql-core==2.3.1       # via graphene, graphql-relay
graphql-relay==2.0.1      # via graphene
grpcio==1.28.1            # via tensorboard, tensorflow
gunicorn==20.0.4          # via -r requirements.in
h11==0.9.0                # via uvicorn
h5py==2.10.0              # via keras, keras-applications
httptools==0.1.1          # via uvicorn
idna==2.9                 # via email-validator, requests
itsdangerous==1.1.0       # via fastapi
jinja2==2.11.1            # via fastapi
keras-applications==1.0.8  # via keras, tensorflow
keras-preprocessing==1.1.0  # via keras, tensorflow
keras==2.3.1              # via -r requirements.in
mako==1.1.2               # via alembic
markdown==3.2.1           # via tensorboard
markupsafe==1.1.1         # via jinja2, mako
networkx==2.5             # via -r requirements.in
numpy==1.18.2             # via h5py, keras, keras-applications, keras-preprocessing, opencv-contrib-python-headless, scipy, tensorboard, tensorflow
opencv-contrib-python-headless==4.4.0.46  # via -r requirements.in
orjson==2.6.3             # via fastapi
passlib[bcrypt]==1.7.2    # via -r requirements.in
pillow==7.1.1             # via -r requirements.in
promise==2.3              # via graphql-core, graphql-relay
protobuf==3.11.3          # via tensorboard, tensorflow
psycopg2==2.8.6           # via -r requirements.in
pycparser==2.20           # via cffi
pydantic==1.4             # via fastapi
pyjwt==1.7.1              # via -r requirements.in
python-dateutil==2.8.1    # via alembic
python-dotenv==0.12.0     # via -r requirements.in
python-editor==1.0.4      # via alembic
python-multipart==0.0.5   # via fastapi
pyyaml==5.3.1             # via fastapi, keras
requests==2.23.0          # via fastapi
rx==1.6.1                 # via graphql-core
scipy==1.4.1              # via keras
six==1.14.0               # via absl-py, bcrypt, google-pasta, graphene, graphql-core, graphql-relay, grpcio, h5py, keras, keras-preprocessing, protobuf, python-dateutil, python-multipart, tensorboard, tensorflow
sqlalchemy==1.3.15        # via -r requirements.in, alembic
starlette==0.13.2         # via fastapi
tensorboard==1.14.0       # via tensorflow
tensorflow-estimator==1.14.0  # via tensorflow
tensorflow==1.14.0        # via -r requirements.in
termcolor==1.1.0          # via tensorflow
ujson==2.0.3              # via fastapi
urllib3==1.25.8           # via requests
uvicorn==0.11.3           # via fastapi
uvloop==0.14.0            # via uvicorn
websockets==8.1           # via uvicorn
werkzeug==1.0.1           # via tensorboard
wheel==0.34.2             # via tensorboard, tensorflow
wrapt==1.12.1             # via tensorflow
exponent_server_sdk      
httpx==0.17.1

# The following packages are considered to be unsafe in a requirements file:
# setuptools